refactor: Replace 'ctx' with 'authContext' in session management imports and create new authContext file
