+++ mobile/app/_layout.tsx
@@ -1,6 +1,6 @@
 import { Stack } from "expo-router";
 import { PaperProvider } from "react-native-paper";
-import { SessionProvider } from "./context/ctx";
+import { SessionProvider } from "./context/authContext";
 import React from "react";
 
 export default function Layout() {
