+++ mobile/app/components/auth/LoginForm.tsx
@@ -1,6 +1,6 @@
 import { AppButton } from "../buttons/AppButton";
 import React from "react";
-import { useSession } from "@/app/context/ctx";
+import { useSession } from "@/app/context/authContext";
 import { FormProvider, useForm } from "react-hook-form";
 import { AppTextField } from "../inputs/AppTextField";
